requirements.txt
Pillow==9.3.0
Django==4.1.3
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
PyYAML==6.0
requests==2.28.2
django-filter==22.1

# transliterate==1.10.2
# django-debug-toolbar==3.8.1
